fix(platform): add u64 cast to windows tempfile retry delay
Fix type mismatch where usize (attempts) was passed to Duration::from_millis()
which expects u64. This resolves compilation errors on 32-bit targets.

Lines 264 and 304 in src/platform.rs now cast (5 * attempts) as u64.

+++ src/platform.rs
@@ -261,7 +261,7 @@ impl NamedTempFileExt for NamedTempFile {
                                 error: err,
                             });
                         }
-                        std::thread::sleep(std::time::Duration::from_millis(5 * attempts));
+                        std::thread::sleep(std::time::Duration::from_millis((5 * attempts) as u64));
                         continue;
                     }
                     return Err(tempfile::PersistError {
@@ -301,7 +301,7 @@ impl TempPathExt for tempfile::TempPath {
                                 error: err,
                             });
                         }
-                        std::thread::sleep(std::time::Duration::from_millis(5 * attempts));
+                        std::thread::sleep(std::time::Duration::from_millis((5 * attempts) as u64));
                         continue;
                     }
                     return Err(tempfile::PathPersistError {
